requirements.txt
# Core data manipulation and scientific computing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Configuration and data generation
pyyaml>=6.0
faker>=18.0.0

# Progress tracking
tqdm>=4.65.0

# Data validation (for future phases)
pytest>=7.3.0

# Date handling
python-dateutil>=2.8.0
